Move query create functions to the neo4j ontology package
