+++ concepts/concepts_service.go
@@ -5,7 +5,6 @@ import (
 	"errors"
 	"fmt"
 	"strconv"
-	"time"
 
 	cmneo4j "github.com/Financial-Times/cm-neo4j-driver"
 	logger "github.com/Financial-Times/go-logger/v2"
@@ -18,7 +17,6 @@ import (
 )
 
 const (
-	iso8601DateOnly = "2006-01-02"
 	//Event types
 	UpdatedEvent = "CONCEPT_UPDATED"
 	AddedEvent   = "CONCORDANCE_ADDED"
@@ -231,8 +229,9 @@ func (s *ConceptService) Write(thing interface{}, transID string) (interface{}, 
 					//TODO fix this up at some point to do it properly?
 					concept.Hash = "0"
 
-					canonical := sourceToCanonical(concept)
-					unconcordQuery := s.writeCanonicalNodeForUnconcordedConcepts(canonical, concept.UUID)
+					canonical := neo4j.SourceToCanonical(concept)
+					unconcordQuery := neo4j.WriteCanonicalNodeForUnconcordedConcepts(canonical, concept.UUID)
+					s.log.WithTransactionID(transID).WithUUID(concept.UUID).Warn("Creating prefUUID node for unconcorded concept")
 					queryBatch = append(queryBatch, unconcordQuery)
 
 					//We will need to send a notification of ids that have been removed from current concordance
@@ -274,7 +273,8 @@ func (s *ConceptService) Write(thing interface{}, transID string) (interface{}, 
 		queryBatch = append(queryBatch, query)
 	}
 	aggregatedConceptToWrite.AggregatedHash = hashAsString
-	queryBatch = populateConceptQueries(queryBatch, aggregatedConceptToWrite)
+	writeQueries := neo4j.WriteCanonicalConceptQueries(aggregatedConceptToWrite)
+	queryBatch = append(queryBatch, writeQueries...)
 
 	updateRecord.UpdatedIds = updatedUUIDList
 	updateRecord.ChangedRecords = append(updateRecord.ChangedRecords, Event{
@@ -406,18 +406,6 @@ func (s *ConceptService) formatError(field, uuid, transID string) string {
 	err := errors.New("invalid request, no " + field + " has been supplied")
 	s.log.WithError(err).WithTransactionID(transID).WithUUID(uuid).Error("Validation of payload failed")
 	return err.Error()
-}
-
-func filterIdsThatAreUniqueToFirstMap(firstMapConcepts map[string]string, secondMapConcepts map[string]string) map[string]string {
-	//Loop through both lists to find id which is present in first list but not in the second
-	filteredMap := make(map[string]string)
-
-	for conceptID := range firstMapConcepts {
-		if _, ok := secondMapConcepts[conceptID]; !ok {
-			filteredMap[conceptID] = firstMapConcepts[conceptID]
-		}
-	}
-	return filteredMap
 }
 
 // Handle new source nodes that have been added to current concordance
@@ -576,17 +564,6 @@ func (s *ConceptService) handleTransferConcordance(conceptData map[string]string
 	return deleteLonePrefUUIDQueries, nil
 }
 
-//Clean up canonical nodes of a concept that has become a source of current concept
-func deleteLonePrefUUID(prefUUID string) *cmneo4j.Query {
-	equivQuery := &cmneo4j.Query{
-		Cypher: `MATCH (t:Thing {prefUUID:$id}) DETACH DELETE t`,
-		Params: map[string]interface{}{
-			"id": prefUUID,
-		},
-	}
-	return equivQuery
-}
-
 //Clear down current concept node
 func (s *ConceptService) clearDownExistingNodes(ac ontology.NewAggregatedConcept) []*cmneo4j.Query {
 	acUUID := ac.PrefUUID
@@ -619,208 +596,15 @@ func (s *ConceptService) clearDownExistingNodes(ac ontology.NewAggregatedConcept
 	return queryBatch
 }
 
-//Curate all queries to populate concept nodes
-func populateConceptQueries(queryBatch []*cmneo4j.Query, aggregatedConcept ontology.NewAggregatedConcept) []*cmneo4j.Query {
-	queryBatch = append(queryBatch, createCanonicalNodeQueries(aggregatedConcept, aggregatedConcept.PrefUUID)...)
-
-	for _, sourceConcept := range aggregatedConcept.SourceRepresentations {
-		queryBatch = append(queryBatch, createNodeQueries(sourceConcept, sourceConcept.UUID)...)
-		queryBatch = append(queryBatch, createEquivalentToQueries(sourceConcept, aggregatedConcept)...)
-
-		for _, rel := range sourceConcept.Relationships {
-			relCfg, ok := ontology.GetConfig().Relationships[rel.Label]
-			if !ok {
-				continue
-			}
-			queryBatch = append(queryBatch, createRelQuery(sourceConcept.UUID, rel, relCfg))
-		}
-	}
-
-	return queryBatch
-}
-
-func filterSlice(a []string) []string {
-	r := []string{}
-	for _, str := range a {
-		if str != "" {
-			r = append(r, str)
-		}
-	}
-	if len(r) == 0 {
-		return nil
-	}
-
-	return r
-}
-
-func createEquivalentToQueries(sourceConcept ontology.NewConcept, aggregatedConcept ontology.NewAggregatedConcept) []*cmneo4j.Query {
-	var queryBatch []*cmneo4j.Query
+//Clean up canonical nodes of a concept that has become a source of current concept
+func deleteLonePrefUUID(prefUUID string) *cmneo4j.Query {
 	equivQuery := &cmneo4j.Query{
-		Cypher: `MATCH (t:Thing {uuid:$uuid}), (c:Thing {prefUUID:$prefUUID})
-						MERGE (t)-[:EQUIVALENT_TO]->(c)`,
+		Cypher: `MATCH (t:Thing {prefUUID:$id}) DETACH DELETE t`,
 		Params: map[string]interface{}{
-			"uuid":     sourceConcept.UUID,
-			"prefUUID": aggregatedConcept.PrefUUID,
+			"id": prefUUID,
 		},
 	}
-
-	queryBatch = append(queryBatch, equivQuery)
-	return queryBatch
-}
-
-func createCanonicalNodeQueries(canonical ontology.NewAggregatedConcept, prefUUID string) []*cmneo4j.Query {
-	var queryBatch []*cmneo4j.Query
-	var createConceptQuery *cmneo4j.Query
-
-	allProps := setCanonicalProps(canonical, prefUUID)
-	createConceptQuery = &cmneo4j.Query{
-		Cypher: fmt.Sprintf(`MERGE (n:Thing {prefUUID: $prefUUID})
-								set n=$allprops
-								set n :%s`, getAllLabels(canonical.Type)),
-		Params: map[string]interface{}{
-			"prefUUID": prefUUID,
-			"allprops": allProps,
-		},
-	}
-
-	queryBatch = append(queryBatch, createConceptQuery)
-	return queryBatch
-}
-
-func createNodeQueries(concept ontology.NewConcept, uuid string) []*cmneo4j.Query {
-	var queryBatch []*cmneo4j.Query
-	var createConceptQuery *cmneo4j.Query
-
-	allProps := setProps(concept, uuid)
-	createConceptQuery = &cmneo4j.Query{
-		Cypher: fmt.Sprintf(`MERGE (n:Thing {uuid: $uuid})
-											set n=$allprops
-											set n :%s`, getAllLabels(concept.Type)),
-		Params: map[string]interface{}{
-			"uuid":     uuid,
-			"allprops": allProps,
-		},
-	}
-
-	if concept.IssuedBy != "" {
-		// Issued By needs a specific handling. That is why it is not in the config
-		// But we still want to use createRelQuery, so we create dummy relationship and config
-		issuedByCfg := ontology.RelationshipConfig{
-			ConceptField: "issuedBy",
-			OneToOne:     true,
-			NeoCreate:    true,
-		}
-		issuedByRel := ontology.Relationship{
-			UUID:       concept.IssuedBy,
-			Label:      "ISSUED_BY",
-			Properties: nil,
-		}
-		queryBatch = append(queryBatch, createRelQuery(concept.UUID, issuedByRel, issuedByCfg))
-	}
-
-	queryBatch = append(queryBatch, createConceptQuery)
-	return queryBatch
-}
-
-// createRelQueries creates relationships Cypher queries for concepts
-func createRelQuery(sourceUUID string, rel ontology.Relationship, cfg ontology.RelationshipConfig) *cmneo4j.Query {
-	const createMissing = `
-		MERGE (thing:Thing {uuid: $uuid})
-		MERGE (other:Thing {uuid: $id})
-		MERGE (thing)-[rel:%s]->(other)
-	`
-
-	const matchExisting = `
-		MATCH (concept:Concept {uuid: $uuid})
-		MERGE (other:Thing {uuid: $id})
-		MERGE (concept)-[rel:%s]->(other)	
-	`
-
-	cypherStatement := matchExisting
-	if cfg.NeoCreate {
-		cypherStatement = createMissing
-	}
-
-	params := map[string]interface{}{
-		"uuid": sourceUUID,
-		"id":   rel.UUID,
-	}
-	if cfg.Properties != nil {
-		cypherStatement += `	SET rel=$relProps`
-		params["relProps"] = setupRelProps(rel, cfg)
-	}
-
-	return &cmneo4j.Query{
-		Cypher: fmt.Sprintf(cypherStatement, rel.Label),
-		Params: params,
-	}
-}
-
-func setupRelProps(rel ontology.Relationship, cfg ontology.RelationshipConfig) map[string]interface{} {
-	props := map[string]interface{}{}
-	for label, t := range cfg.Properties {
-		val := rel.Properties[label]
-		props[label] = val
-		if val != nil && t == ontology.PropertyTypeDate {
-			str, ok := rel.Properties[label].(string)
-			if !ok {
-				continue
-			}
-			unixTime := getEpoch(str)
-			// in the old times we skipped unix timestamps with valuse less or equal 0
-			if unixTime <= 0 {
-				continue
-			}
-			props[label+"Epoch"] = unixTime
-		}
-	}
-	return props
-}
-
-func getEpoch(t string) int64 {
-	if t == "" {
-		return 0
-	}
-
-	tt, err := time.Parse(iso8601DateOnly, t)
-	if err != nil {
-		return 0
-	}
-	unixTime := tt.Unix()
-	if unixTime < 0 {
-		return 0
-	}
-	return unixTime
-}
-
-//Create canonical node for any concepts that were removed from a concordance and thus would become lone
-func (s *ConceptService) writeCanonicalNodeForUnconcordedConcepts(canonical ontology.NewAggregatedConcept, prefUUID string) *cmneo4j.Query {
-	allProps := setCanonicalProps(canonical, prefUUID)
-	s.log.WithField("UUID", prefUUID).Warn("Creating prefUUID node for unconcorded concept")
-	createCanonicalNodeQuery := &cmneo4j.Query{
-		Cypher: fmt.Sprintf(`
-					MATCH (t:Thing{uuid:$prefUUID})
-					MERGE (n:Thing {prefUUID: $prefUUID})
-					MERGE (n)<-[:EQUIVALENT_TO]-(t)
-					set n=$allprops
-					set n :%s`, getAllLabels(canonical.Type)),
-		Params: map[string]interface{}{
-			"prefUUID": prefUUID,
-			"allprops": allProps,
-		},
-	}
-	return createCanonicalNodeQuery
-}
-
-//return all concept labels
-func getAllLabels(conceptType string) string {
-	labels := conceptType
-	parentType := mapper.ParentType(conceptType)
-	for parentType != "" {
-		labels += ":" + parentType
-		parentType = mapper.ParentType(parentType)
-	}
-	return labels
+	return equivQuery
 }
 
 //extract uuids of the source concepts
@@ -832,72 +616,16 @@ func getSourceData(sourceConcepts []ontology.NewConcept) map[string]string {
 	return conceptData
 }
 
-//This function dictates which properties will be actually
-//written in neo for source nodes.
-func setProps(source ontology.NewConcept, uuid string) map[string]interface{} {
-	nodeProps := map[string]interface{}{}
-	nodeProps["lastModifiedEpoch"] = time.Now().Unix()
+func filterIdsThatAreUniqueToFirstMap(firstMapConcepts map[string]string, secondMapConcepts map[string]string) map[string]string {
+	//Loop through both lists to find id which is present in first list but not in the second
+	filteredMap := make(map[string]string)
 
-	if source.PrefLabel != "" {
-		nodeProps["prefLabel"] = source.PrefLabel
-	}
-
-	if source.FigiCode != "" {
-		nodeProps["figiCode"] = source.FigiCode
-	}
-
-	if source.IsDeprecated {
-		nodeProps["isDeprecated"] = true
-	}
-
-	nodeProps["uuid"] = uuid
-	nodeProps["authority"] = source.Authority
-	nodeProps["authorityValue"] = source.AuthorityValue
-
-	return nodeProps
-}
-
-//This function dictates which properties will be actually
-//written in neo for canonical nodes.
-func setCanonicalProps(canonical ontology.NewAggregatedConcept, prefUUID string) map[string]interface{} {
-	nodeProps := map[string]interface{}{}
-
-	ontologyCfg := ontology.GetConfig()
-	for field, propCfg := range ontologyCfg.Fields {
-		if val, ok := canonical.GetPropertyValue(field); ok {
-			if !ontologyCfg.IsPropValueValid(field, val) {
-				continue
-			}
-
-			nodeProps[propCfg.NeoProp] = val
+	for conceptID := range firstMapConcepts {
+		if _, ok := secondMapConcepts[conceptID]; !ok {
+			filteredMap[conceptID] = firstMapConcepts[conceptID]
 		}
 	}
-
-	nodeProps["lastModifiedEpoch"] = time.Now().Unix()
-
-	if canonical.PrefLabel != "" {
-		nodeProps["prefLabel"] = canonical.PrefLabel
-	}
-
-	if canonical.FigiCode != "" {
-		nodeProps["figiCode"] = canonical.FigiCode
-	}
-
-	if canonical.IsDeprecated {
-		nodeProps["isDeprecated"] = true
-	}
-
-	nodeProps["prefUUID"] = prefUUID
-	nodeProps["aggregateHash"] = canonical.AggregatedHash
-
-	if canonical.InceptionDate != "" {
-		nodeProps["inceptionDate"] = canonical.InceptionDate
-	}
-	if canonical.TerminationDate != "" {
-		nodeProps["terminationDate"] = canonical.TerminationDate
-	}
-
-	return nodeProps
+	return filteredMap
 }
 
 //DecodeJSON - decode json
@@ -953,34 +681,4 @@ func stringInArr(searchFor string, values []string) bool {
 		}
 	}
 	return false
-}
-
-func sourceToCanonical(source ontology.NewConcept) ontology.NewAggregatedConcept {
-	var inceptionDate string
-	var terminationDate string
-	for _, r := range source.Relationships {
-		if r.Label != "HAS_ROLE" {
-			continue
-		}
-		if v, ok := r.Properties["inceptionDate"]; ok {
-			if s, ok := v.(string); ok {
-				inceptionDate = s
-			}
-		}
-		if v, ok := r.Properties["terminationDate"]; ok {
-			if s, ok := v.(string); ok {
-				terminationDate = s
-			}
-		}
-		break
-	}
-	return ontology.NewAggregatedConcept{
-		AggregatedHash:  source.Hash,
-		InceptionDate:   inceptionDate,
-		IssuedBy:        source.IssuedBy,
-		PrefLabel:       source.PrefLabel,
-		TerminationDate: terminationDate,
-		Type:            source.Type,
-		IsDeprecated:    source.IsDeprecated,
-	}
 }
