+++ ontology/neo4j/testdata/brand-with-multiple-implied-by.json
@@ -1,0 +1,24 @@
+{  
+    "prefUUID":"cce1bc63-3717-4ae6-9399-88dab5966815",
+    "prefLabel":"Pref Label",
+    "type":"Brand",
+    "sourceRepresentations":[  
+        {  
+            "uuid":"cce1bc63-3717-4ae6-9399-88dab5966815",
+            "prefLabel":"Pref Label",
+            "type":"Brand",
+            "authority":"Smartlogic",
+            "authorityValue":"1234",
+            "aliases":[  
+                "brandLabel",
+                "secondBrandLabel",
+                "anotherBrandOne",
+                "whyNotBrand"
+            ],
+            "impliedByUUIDs":[  
+                "740c604b-8d97-443e-be70-33de6f1d6e67",
+                "b5d7c6b5-db7d-4bce-9d6a-f62195571f92"
+            ]
+        }
+    ]
+}
