+++ ontology/neo4j/testdata/concept-queries-country-of-naics-rels.golden
