+++ ontology/neo4j/testdata/concept-queries-default-source.golden
