+++ ontology/neo4j/testdata/concept-queries-has-broader-rel.golden
