+++ ontology/neo4j/testdata/concept-queries-has-focus-rel.golden
