+++ ontology/neo4j/testdata/concept-queries-implied-by-rel.golden
