+++ ontology/neo4j/testdata/concept-queries-is-related-to-rel.golden
