+++ ontology/neo4j/testdata/concept-queries-membership-rels.golden
