+++ ontology/neo4j/testdata/concept-queries-sub-organisation-of-rel.golden
