+++ ontology/neo4j/testdata/concept-with-multiple-has-focus.json
@@ -1,0 +1,25 @@
+{  
+    "prefUUID":"a39a4558-f562-4dca-8774-000246e6eebe",
+    "prefLabel":"Pref Label",
+    "type":"Organisation",
+    "sourceRepresentations":[  
+        {  
+            "uuid":"a39a4558-f562-4dca-8774-000246e6eebe",
+            "prefLabel":"Pref Label",
+            "type":"Organisation",
+            "authority":"Smartlogic",
+            "authorityValue":"1234",
+            "aliases":[  
+                "brandLabel",
+                "secondBrandLabel",
+                "anotherBrandOne",
+                "whyNotBrand"
+            ],
+            "hasFocusUUIDs":[  
+                "2e7429bd-7a84-41cb-a619-2c702893e359",
+                "740c604b-8d97-443e-be70-33de6f1d6e67",
+                "c28fa0b4-4245-11e8-842f-0ed5f89f718b"
+            ]
+        }
+    ]
+}
