+++ ontology/neo4j/testdata/concept-with-multiple-superseded-by.json
@@ -1,0 +1,24 @@
+{  
+    "prefUUID":"bbc4f575-edb3-4f51-92f0-5ce6c708d1ea",
+    "prefLabel":"Pref Label",
+    "type":"Section",
+    "sourceRepresentations":[  
+        {  
+            "uuid":"bbc4f575-edb3-4f51-92f0-5ce6c708d1ea",
+            "prefLabel":"Pref Label",
+            "type":"Section",
+            "authority":"Smartlogic",
+            "authorityValue":"1234",
+            "aliases":[  
+                "oneLabel",
+                "secondLabel",
+                "anotherOne",
+                "whyNot"
+            ],
+            "supersededByUUIDs":[  
+                "1a96ee7a-a4af-3a56-852c-60420b0b8da6",
+                "b5d7c6b5-db7d-4bce-9d6a-f62195571f92"
+            ]
+        }
+    ]
+}
